pyproject.toml
[project]
name = "veo3-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.21.0",
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "boto3>=1.40.4",
    "celery>=5.5.3",
    "elastic-apm>=6.24.0",
    "fastapi[standard]>=0.116.1",
    "flower>=2.0.1",
    "kibana>=0.7",
    "openai>=1.99.9",
    "pydantic-settings>=2.10.1",
    "redis>=6.4.0",
    "scalar-fastapi>=1.3.0",
    "sqlalchemy>=2.0.42",
    "yookassa>=3.7.0",
]
